project hierarchy changed, problem with databasemanager deleted sorted
