+++ TDR/Assets/Scripts/Cloud/LoginManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ccdf26f55bc81ed4b89ec5a301f63ecc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
