+++ TDR/Assets/Scripts/Cloud/WaitingRoomManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6b1136cce2b48424d89b57632a7a07fa
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
